+++ lib/leanplum_api/data_export_api.rb
@@ -59,6 +59,13 @@ module LeanplumApi
       end
     end
 
+    # See leanplum docs.
+    # The segment syntax is identical to that produced by the "Insert Value" feature on the dashboard.
+    # Examples: 'Country = "US"', '{Country = "US"} and {App version = 1}'.
+    def export_users(ab_test_id = nil, segment = nil)
+      data_export_connection.get(action: 'exportUsers', segment: segment, ab_test_id: ab_test_id).body['response'].first['jobId']
+    end
+
     def wait_for_export_job(job_id, polling_interval = 60)
       while get_export_results(job_id)[:state] != EXPORT_FINISHED
         LeanplumApi.configuration.logger.debug("Polling job #{job_id}: #{get_export_results(job_id)}")