+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * **BREAKING CHANGE** - `track_multi` method signature has changed.
 * **BREAKING CHANGE** - Data export methods moved to the `DataExportAPI` class
 * *New Feature* - Added support for setDeviceAttributes (thanks @beingmattlevy)
+* Fix `export_users`
 
 ## 3.1.0
 * Add the ability to send events as `userAttributes` properties